+++ libvec2/srcs/libvec2.c
@@ -1,6 +1,6 @@
 #include "libvec2.h"
 
-t_vec2	vec2_init(float x, float y)
+t_vec2	vec2_init(double x, double y)
 {
 	return ((t_vec2){x, y});
 }
@@ -10,19 +10,19 @@ t_vec2	vec2_add(t_vec2 a, t_vec2 b)
 	return ((t_vec2){a.x + b.x, a.y + b.y});
 }
 
-float	vec2_dot(t_vec2 a, t_vec2 b)
+double	vec2_dot(t_vec2 a, t_vec2 b)
 {
 	return (a.x * b.x + a.y * b.y);
 }
 
-float	vec2_norme(t_vec2 a)
+double	vec2_norme(t_vec2 a)
 {
 	return (sqrt(a.x * a.x + a.y * a.y));
 }
 
 t_vec2	vec2_normalize(t_vec2 a)
 {
-	float norm;
+	double norm;
 
 	norm = vec2_norme(a);
 	return ((t_vec2){a.x / norm, a.y / norm});
